Move all clang related code into clang_ast_to_asr.h
Create temporary clang_ast_to_asr2.h
